refactoring and folder changes to accomodate SPA with ties
