+++ src/algmatch/stableMatchings/studentProjectAllocation/noTies/DATA_FORMAT_GUIDELINE.md
