+++ src/algmatch/stableMatchings/studentProjectAllocation/noTies/spaAbstract.py
@@ -4,7 +4,7 @@ Student Project Allocation - Abstract class
 
 import os
 
-from algmatch.stableMatchings.studentProjectAllocation.spaPreferenceInstance import SPAPreferenceInstance
+from algmatch.stableMatchings.studentProjectAllocation.noTies.spaPreferenceInstance import SPAPreferenceInstance
 
 
 class SPAAbstract:
